added 'PPM', page-per-minute statistics
git-svn-id: https://svn.berlios.de/svnroot/repos/yacy/trunk@405 6c8d7289-2bf4-0310-a012-ef5d649a1542

+++ source/de/anomic/plasma/plasmaHTCache.java
@@ -205,7 +205,7 @@ public final class plasmaHTCache {
         File f;
         while (currCacheSize > maxCacheSize) {
             f = (File) cacheAge.remove(cacheAge.firstKey());
-            if (f.exists()) {
+            if ((f != null) && (f.exists())) {
                 currCacheSize -= f.length();
                 if (f.delete()) {
                     log.logInfo("DELETED OLD CACHE : " + f.toString());
